add feature to input xslt directory to test new transforms

+++ teiToJSON.js
@@ -3,19 +3,34 @@ const exec = util.promisify(require('child_process').exec);
 const fs = require('fs/promises');
 const path = require('path');
 
-async function processDataWithDocker(teiDirectory, isSudoDocker) {
+async function processDataWithDocker(teiDirectory, xsltDirectory='', isSudoDocker) {
 
     const sudoDockerString = isSudoDocker ? 'sudo docker' : 'docker';
+    // to manually shell into docker container, run:
+    // sudo docker run -it --rm cudl-xslt:0.0.5
 
 
     try {
         // Create a symbolic link to the data directory
         const { stdout: linkOutput } = await exec(`ln -s ${teiDirectory} data`);
-        console.log('Symbolic link created:', linkOutput.trim());
+        console.log('Symbolic link created for data directory:', linkOutput.trim());
 
-        // Run the ant build command inside the Docker container with local volume mount
+        if(xsltDirectory === ''){
+            //No XSLT directory specified
+            // Run the ant build command inside the Docker container with local volume mount
         const { stdout: antOutput } = await exec(`${sudoDockerString} run --rm -v ${__dirname}/data:/opt/data -v ${__dirname}/json:/opt/json shenukacj/cudl-xslt:0.0.5 ant -buildfile ./bin/build.xml "json"`);
         console.log('Ant command executed:', antOutput.trim());
+        }
+
+        else{
+            // else volume mount to the xslt directory as well
+            const { stdout: linkOutput } = await exec(`ln -s ${xsltDirectory} xslt`);
+            console.log('Symbolic link created for xslt directory:', linkOutput.trim());
+            const { stdout: antOutput } = await exec(`${sudoDockerString} run --rm -v ${__dirname}/data:/opt/data -v ${__dirname}/json:/opt/json -v ${__dirname}/xslt:/opt/xslt shenukacj/cudl-xslt:0.0.5 ant -buildfile ./bin/build.xml "json"`);
+        console.log('Ant command executed:', antOutput.trim());
+        }
+
+        
 
         // Remove the data directory in the Docker container
         const { stdout: removeOutput } = await exec('rm -r data');
@@ -58,8 +73,18 @@ async function readJsonFiles(filePath) {
 }
 
 // Example usage
-processDataWithDocker('/home/linux/Documents/github/dl-data-samples/source-data/data/items/data/tei/MS-TEST-ITEM-00002/', true)
+// processDataWithDocker('./example-data-directory', true)
+//     .then((jsonData) => {
+//         console.log('JSON data:', jsonData);
+//     })
+//     .catch((err) => {
+//         console.error('Error:', err);
+//     });
+
+// Example data with XSLT directory added
+processDataWithDocker('./example-data-directory', './example-xslt-directory', true)
     .then((jsonData) => {
+        console.log('process with XSLT directory input')
         console.log('JSON data:', jsonData);
     })
     .catch((err) => {
